% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_this_file}
\alias{check_this_file}
\title{Check file existence and optionally create directory}
\usage{
check_this_file(file_path, mode = c("write", "read"), force = FALSE)
}
\arguments{
\item{file_path}{The path to the file to be checked. If NULL, an error is raised.}

\item{mode}{The mode of operation. Choose from "write" or "read".}

\item{force}{Logical indicating whether to force the operation (default is FALSE).}
}
\description{
This function checks whether a file exists and, depending on the mode, performs additional checks.
If the mode is set to "write", it checks whether the file exists. If it exists and force is FALSE,
an error is raised. If force is TRUE, the existing file is overwritten or created if it doesn't exist.
If the mode is set to "read", it checks whether the file exists. If it doesn't, an error is raised.
}
\examples{
\dontrun{
# Example with a temporary file for writing
temp_file_write <- tempfile(fileext = ".txt")
check_this_file(temp_file_write, mode = "write")

# Example creating a file with file.create() and then checking it for writing
temp_file_create <- tempfile(fileext = ".txt")
file.create(temp_file_create)
check_this_file(temp_file_create, mode = "write", force = TRUE)

# Example checking a non-existing file for reading
temp_file_read <- tempfile(fileext = ".txt")
check_this_file(temp_file_read, mode = "read")
}

}
