% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_this_var}
\alias{check_this_var}
\title{Check variable type.}
\usage{
check_this_var(
  x,
  type = c("char", "num", "bool", "int"),
  x_name = deparse(substitute(x)),
  null_accepted = FALSE,
  calling_function = NULL
)
}
\arguments{
\item{x}{The variable to be checked.}

\item{type}{The expected type. Choose from "char", "num", "bool", or "int".}

\item{x_name}{The name of the variable as a string.}

\item{null_accepted}{Whether NULL values are accepted.}

\item{calling_function}{The name of the calling function (automatically obtained if envnames is available).}
}
\description{
This function checks the type of a variable and raises an error if it does not match the expected type.
}
\examples{
\dontrun{
# Example checking a character variable
check_this_var(x = "hello", type = "char")

# Example checking a numeric variable
check_this_var(x = 42, type = "num")
}

}
