% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{clustermole_subset}
\alias{clustermole_subset}
\title{The clustermole_subset dataset}
\format{
A list of size 28.
}
\usage{
data(clustermole_subset)
}
\description{
Few markers from the clustermole package This dataset is used for testing purposes.
}
\note{
This dataset was produced using the clustermole package:
m <- clustermole_markers(species = "hs")
clustermole_subset <- split(m$gene[m$organ == "Blood" & m$species == "Human"],
                            m$celltype[m$organ == "Blood" & m$species == "Human"])
}
\examples{
data(clustermole_subset)

}
\keyword{datasets}
