% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors_for_gradient}
\alias{colors_for_gradient}
\title{Generate a vector of colors for a gradient}
\usage{
colors_for_gradient(palette = "Je1")
}
\arguments{
\item{palette}{A character string specifying the palette to use.}
}
\value{
A character vector of color codes.
}
\description{
This function generates a vector of colors for a gradient, given
a specified palette name.
}
\examples{
colors_for_gradient()
colors_for_gradient(palette = "Ju1")

}
