% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{create_rand_str}
\alias{create_rand_str}
\title{Generate a random string of letters and numbers}
\usage{
create_rand_str()
}
\value{
A character string of length 10, consisting of random letters
  and numbers.
}
\description{
This function generates a random string of 10 characters, consisting
of 3 uppercase letters, 4 digits, and 3 lowercase letters.
}
\examples{
create_rand_str()
}
