% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{ident_pbmc3k_medium}
\alias{ident_pbmc3k_medium}
\title{The ident_pbmc3k_medium dataset}
\format{
A named vector of size 361.
}
\usage{
data(ident_pbmc3k_medium)
}
\description{
Cell identity (ident) from the pbmc3k_medium dataset. This dataset is used for testing purposes.
}
\examples{
data(ident_pbmc3k_medium)

}
\keyword{datasets}
