% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{load_example}
\alias{load_example}
\title{Load an example dataset}
\usage{
load_example(path, envir = parent.frame())
}
\arguments{
\item{path}{Path returned by [get_example()].}

\item{envir}{The environment into which the dataset is loaded.}
}
\value{
Invisibly, a character vector of the names of objects loaded.
}
\description{
Loads an example dataset downloaded with [get_example()] into an environment.
}
\examples{
\dontrun{
set_verbosity(3)
load_example(get_example("22745973/files/ENCFF119BYM_H3K36me3_chr1"))
}

}
