% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity.R
\name{print_msg}
\alias{print_msg}
\title{Print a message based on the level of verbosity.}
\usage{
print_msg(..., msg_type = c("INFO", "DEBUG", "WARNING", "STOP", "ERROR"))
}
\arguments{
\item{...}{The message components to be printed (pasted together).}

\item{msg_type}{The type of message, one of "INFO", "DEBUG", "WARNING",
"STOP", or "ERROR" (ERROR is translated to STOP).}
}
\value{
NULL, invisibly.
}
\description{
Print a message based on the level of verbosity.
}
\examples{
# Set verbosity level to 1
library(dplab) 
set_verbosity(1)
print_msg("Hello world!", msg_type = "INFO")
set_verbosity(2)
print_msg("Debugging message", msg_type = "DEBUG")
set_verbosity(0)
print_msg("Hello world!", msg_type = "INFO")
print_msg("Debugging message", msg_type = "DEBUG")
}
