% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{rand_string}
\alias{rand_string}
\title{Generate a random string of letters and numbers (alias)}
\usage{
rand_string()
}
\value{
A character string of length 10, consisting of random letters
  and numbers.
}
\description{
Alias for \code{\link{create_rand_str}}. Generates a random string
of 10 characters.
}
\examples{
rand_string()
}
