% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity.R
\name{set_verbosity}
\alias{set_verbosity}
\title{Set the verbosity level.}
\usage{
set_verbosity(verbosity_value)
}
\arguments{
\item{verbosity_value}{A non-negative integer indicating the verbosity level to be set.}
}
\value{
NULL, invisibly.
}
\description{
This function sets the verbosity level which controls the amount of
information that is printed to the console by the \code{\link{print_msg}}
function. The verbosity level can be set to any non-negative integer,
with higher values indicating more detailed output. By default, the
verbosity level is set to 1.
}
\examples{
# Set verbosity level to 2
set_verbosity(2)

# Set verbosity level to 0
set_verbosity(0)

# 0 : No message
# 1 : Display only INFO type message
# 2 : Display both INFO and DEBUG type message
}
