% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{show_methods}
\alias{show_methods}
\title{List the methods for an S4 class}
\usage{
show_methods(class = NULL, where = ".GlobalEnv")
}
\arguments{
\item{class}{The name of the S4 class (character string) or an object
of that class. If \code{NULL}, methods for all classes are shown.}

\item{where}{The environment or package where the class is defined.
Defaults to \code{".GlobalEnv"}. If a package name is supplied,
it is automatically converted to \code{"package:<package>"}.}
}
\value{
A character vector containing the available method names.
}
\description{
List the methods available for an S4 class in a specified environment
or package.
}
\examples{
\dontrun{
# Define an example S4 class in the global environment
methods::setClass(
  "ExampleClass",
  slots = list(
    name = "character",
    values = "numeric"
  )
)

# Define an example S4 method
methods::setMethod(
  "show",
  "ExampleClass",
  function(object) {
    cat("ExampleClass:", object@name, "\n")
  }
)

# List methods for the class
show_methods("ExampleClass")

# Alternatively, pass an object
example_object <- methods::new(
  "ExampleClass",
  name = "test",
  values = c(1, 2, 3)
)

show_methods(example_object)
}

}
