% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{denlab_gg_theme}
\alias{denlab_gg_theme}
\title{Custom ggplot2 Theming for DenLab Packages}
\usage{
denlab_gg_theme()
}
\value{
A ggplot2 theme object with predefined settings for legend, axis, and strip text.
}
\description{
This function defines a custom theme for ggplot2 plots with specific adjustments
to the legend key dimensions, legend text, axis titles, and strip text sizes.
}
\details{
The \code{denlab_gg_theme} function customizes the appearance of ggplot2 plots.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point() +
  denlab_gg_theme()
}

}
