% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{mkdir_p}
\alias{mkdir_p}
\title{Create a directory (including parents) if it doesn't exist}
\usage{
mkdir_p(path)
}
\arguments{
\item{path}{The path to the directory to create (character string).}
}
\value{
NULL, invisibly. Creates the directory as a side effect.
}
\description{
Creates a directory recursively if it doesn't already exist. 
Similar to `mkdir -p` in Unix. Uses \code{\link{print_msg}} to inform the user.
}
\examples{
\dontrun{
# Create a directory and its parents if they don't exist
# print a msg using print_msg() (controlled by set_verbosity())
mkdir_p("/tmp/denlabutils")

# Call again - will not error if already exists
mkdir_p("/tmp/denlabutils")

# Set verbosity level to see DEBUG messages
set_verbosity(2)
mkdir_p("/tmp/denlabutils/subdir")
mkdir_p("/tmp/denlabutils/subdir")
}
}
