% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{reload_pac}
\alias{reload_pac}
\title{Reload a package (used for development).}
\usage{
reload_pac(package_name = "denlabutils")
}
\arguments{
\item{package_name}{The name of the package to reload. Defaults to "denlabutils".}
}
\value{
NULL, invisibly. Prints messages about the reload process.
}
\description{
Detach and reload a package. Useful during package development to test changes
without restarting R. By default, reloads "denlabutils".
}
\examples{
\dontrun{
# Reload denlabutils
reload_pac()

# Reload another package
reload_pac("mypackage")
}

}
